project_03_airflow_custom_docker/requirements.txt
psycopg2-binary
python-dotenv
tqdm
pandas
numpy